delete few checks in the test
