+++ tests/manage/z_cluster/cluster_expansion/test_add_capacity_with_node_restart.py
@@ -6,13 +6,48 @@ from concurrent.futures import ThreadPoolExecutor
 from ocs_ci.framework.pytest_customization.marks import polarion_id
 from ocs_ci.framework.testlib import ignore_leftovers, ManageTest, tier4
 from ocs_ci.framework import config
-from ocs_ci.ocs import constants
+from ocs_ci.ocs import constants, defaults
 from ocs_ci.ocs.ocp import OCP
 from ocs_ci.ocs.resources import pod as pod_helpers
 from ocs_ci.ocs.resources import storage_cluster
 from ocs_ci.utility.utils import ceph_health_check
 from ocs_ci.ocs.node import get_typed_nodes, wait_for_nodes_status
-from tests.manage.z_cluster.cluster_expansion import temp_helper_file
+
+
+# The functions below I took from the branch 'wip-add-capacity-e_e'.
+# When this branch will be merged into master I will use the functions from this branch.
+
+def get_percent_used_capacity():
+    """
+    Function to calculate the percentage of used capacity in a cluster
+
+    Returns:
+        float: The percentage of the used capacity in the cluster
+    """
+    ct_pod = pod_helpers.get_ceph_tools_pod()
+    output = ct_pod.exec_ceph_cmd(ceph_cmd='ceph df')
+    total_used = (output.get('stats').get('total_used_raw_bytes'))
+    total_avail = (output.get('stats').get('total_bytes'))
+    return 100.0 * total_used / total_avail
+
+
+def check_pods_in_running_state(namespace=defaults.ROOK_CLUSTER_NAMESPACE):
+    """
+    checks whether all the pods in a given namespace are in Running state or not
+    Returns:
+        Boolean: True, if all pods in Running state. False, otherwise
+    """
+    ret_val = True
+    list_of_pods = pod_helpers.get_all_pods(namespace)
+    ocp_pod_obj = OCP(kind=constants.POD, namespace=namespace)
+    for p in list_of_pods:
+        # we don't want to compare osd-prepare and canary pods as they get created freshly when an osd need to be added.
+        if "rook-ceph-osd-prepare" not in p.name and "rook-ceph-drain-canary" not in p.name:
+            status = ocp_pod_obj.get_resource(p.name, 'STATUS')
+            if status not in "Running":
+                logging.error(f"The pod {p.name} is in {status} state. Expected = Running")
+                ret_val = False
+    return ret_val
 
 
 @pytest.mark.parametrize(
@@ -53,28 +88,27 @@ class TestAddCapacityNodeRestart(ManageTest):
         in the middle of the process
         """
         logging.info("Condition 1 to start the test is met: storageutilization is completed")
-        used_capacity_after_utilizing = temp_helper_file.get_percent_used_capacity()
+        used_capacity_after_utilizing = get_percent_used_capacity()
         logging.info(f"### used capacity after storageutilization = {used_capacity_after_utilizing}")
 
         seconds_to_wait_after_storageutilization = 60
         logging.info(f"wait {seconds_to_wait_after_storageutilization} seconds before starting the test")
         time.sleep(seconds_to_wait_after_storageutilization)
 
-        assert temp_helper_file.check_pods_in_running_state(), \
-            "Condition 2 to start test failed: one or more OCS pods are not in running state"
+        assert check_pods_in_running_state(), (
+            "Condition 2 to start test failed: one or more OCS pods are not in running state")
 
         node_list = get_typed_nodes(node_type=node_type, num_of_nodes=num_of_nodes)
         assert node_list, "Condition 3 to start test failed: No node to restart"
 
         osd_pods_before = pod_helpers.get_osd_pods()
-        assert len(osd_pods_before) < 9, \
-            "Condition 4 to start test failed: We have maximum of osd's in the cluster"
+        assert len(osd_pods_before) < 9, (
+            "Condition 4 to start test failed: We have maximum of osd's in the cluster")
         logging.info("All start conditions are met!")
 
-        total_space_before_expansion = temp_helper_file.get_total_space()
-        logging.info('total space  before expansion = {}'.format(str(total_space_before_expansion)))
-
         logging.info("Perform some IO operations...")
+        # The IOs here are not complete. When the branch 'wip-add-capacity-e_e' will be merged into master
+        # I will use the functions from this branch.
 
         pvc_objs = multi_pvc_factory(
             size=self.pvc_size, num_of_pvc=self.num_of_pvcs
@@ -91,10 +125,10 @@ class TestAddCapacityNodeRestart(ManageTest):
                 pod_io_task = executor.submit(
                     pod_obj.run_io, storage_type='fs', size=2, jobs=10, io_direction='wo', rate='200m')
 
-        seconds_to_wait_for_io_operations = 90
+        seconds_to_wait_for_io_operations = 60
         logging.info(f"Going to sleep for {seconds_to_wait_for_io_operations} seconds")
         time.sleep(seconds_to_wait_for_io_operations)
-        used_capacity_after_io_operations = temp_helper_file.get_percent_used_capacity()
+        used_capacity_after_io_operations = get_percent_used_capacity()
         logging.info(f"### used capacity after IO = {used_capacity_after_io_operations}")
 
         osd_size = storage_cluster.get_osd_size()
@@ -112,8 +146,11 @@ class TestAddCapacityNodeRestart(ManageTest):
         wait_for_nodes_status(node_names=[node.name for node in node_list])
         logging.info("Finished restarting the node list")
 
-        used_capacity_after_io_operations = temp_helper_file.get_percent_used_capacity()
+        used_capacity_after_io_operations = get_percent_used_capacity()
         logging.info(f"### used capacity after node restart = {used_capacity_after_io_operations}")
+
+        # The exit criteria verification conditions here are not complete. When the branch
+        # 'wip-add-capacity-e_e' will be merged into master I will use the functions from this branch.
 
         pod = OCP(
             kind=constants.POD, namespace=config.ENV_DATA['cluster_namespace']
@@ -124,14 +161,6 @@ class TestAddCapacityNodeRestart(ManageTest):
             selector='app=rook-ceph-osd',
             resource_count=result * 3
         )
-
-        assert temp_helper_file.check_osd_pods_after_expansion(osd_pods_before), (
-            "number of osd pods is not as expected"
-        )
-        assert temp_helper_file.check_total_space_after_expansion(total_space_before_expansion), (
-            "Expected capacity mismatch"
-        )
-        assert temp_helper_file.check_osd_tree(), "Incorrect ceph osd tree format"
 
         logging.info("Finished verifying add capacity osd storage with node restart")
         logging.info("Waiting for ceph health check to finished...")
